+++ src/modules/components/odsModulesProvider/index.ts
@@ -1,1 +1,6 @@
-export { OdsModulesProvider, type IOdsModulesProviderProps } from './odsModulesProvider';
+export {
+    OdsModulesProvider,
+    useOdsModulesContext,
+    type IOdsModulesContext,
+    type IOdsModulesProviderProps,
+} from './odsModulesProvider';
